feat: automatically create starter and freebuild groups and do not allow deleting them

+++ src/main/java/de/unknowncity/plots/command/admin/PlotAdminGroupDeleteCommand.java
@@ -48,6 +48,12 @@ public class PlotAdminGroupDeleteCommand extends SubCommand {
         var player = (Player) commandContext.sender();
         var group = commandContext.<PlotGroup>get("group");
 
+        if (group.name().equals(plugin.configuration().fb().freeBuildGroup()) ||
+                group.name().equals(plugin.configuration().starterPlotGroup())) {
+            plugin.messenger().sendMessage(player, NodePath.path("command", "plotadmin", "group", "delete", "internal"));
+            return;
+        }
+
         plotService.deletePlotGroup(group.name());
 
         plugin.messenger().sendMessage(player, NodePath.path("command", "plotadmin", "group", "delete", "success"));
